Fix install script

+++ netkit_python/install.py
@@ -10,9 +10,10 @@ install_directory.mkdir(mode=0o775, exist_ok=True)
 
 print("Copying files:")
 for source_file in source_directory.iterdir():
-    destination_file = install_directory / source_file.name
-    print(f"  Copying {source_file} to {destination_file}.")
-    copyfile(source_file, destination_file)
+    if source_file.is_file():
+        destination_file = install_directory / source_file.name
+        print(f"  Copying {source_file} to {destination_file}.")
+        copyfile(source_file, destination_file)
 
 print("Making wrapper scripts:")
 wrapper_directory = common.netkit_home / "bin"
